WIP: refactor library editor to be able to display internal libs too(e.g d.ts)
